Integration of various components - windows awsvpc
Major changes made as part of this commit -
1. Integrated various individual components such as ENI Watcher and CNI
plugins
2. Moved common code from platform specific file and vice versa
3. Created new plugin configuration and invokation logic for Windows
4. Integrated additional flows required for awsvpc mode
