+++ agent/app/agent_capability_unspecified.go
@@ -16,6 +16,7 @@
 package app
 
 import (
+	"errors"
 	"strings"
 
 	"github.com/aws/amazon-ecs-agent/agent/config"
@@ -112,4 +113,9 @@ func (agent *ecsAgent) appendGMSACapabilities(capabilities []*ecs.Attribute) []*
 
 func (agent *ecsAgent) appendEFSVolumePluginCapabilities(capabilities []*ecs.Attribute, pluginCapability string) []*ecs.Attribute {
 	return capabilities
+}
+
+// getTaskENIPluginVersionAttribute for unsupported platform would return an error
+func (agent *ecsAgent) getTaskENIPluginVersionAttribute() (*ecs.Attribute, error) {
+	return nil, errors.New("unsupported platform")
 }
