+++ agent/app/agent_unix.go
@@ -16,7 +16,6 @@
 package app
 
 import (
-	"fmt"
 	"os"
 
 	asmfactory "github.com/aws/amazon-ecs-agent/agent/asm/factory"
@@ -34,7 +33,6 @@ import (
 	"github.com/aws/amazon-ecs-agent/agent/taskresource"
 	cgroup "github.com/aws/amazon-ecs-agent/agent/taskresource/cgroup/control"
 	"github.com/aws/amazon-ecs-agent/agent/utils/ioutilwrapper"
-	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/cihub/seelog"
 	"github.com/pkg/errors"
 )
@@ -97,42 +95,6 @@ func (agent *ecsAgent) initializeTaskENIDependencies(state dockerstate.TaskEngin
 	return nil, false
 }
 
-// setVPCSubnet sets the vpc and subnet ids for the agent by querying the
-// instance metadata service
-func (agent *ecsAgent) setVPCSubnet() (error, bool) {
-	mac, err := agent.ec2MetadataClient.PrimaryENIMAC()
-	if err != nil {
-		return fmt.Errorf("unable to get mac address of instance's primary ENI from instance metadata: %v", err), false
-	}
-
-	vpcID, err := agent.ec2MetadataClient.VPCID(mac)
-	if err != nil {
-		if isInstanceLaunchedInVPC(err) {
-			return fmt.Errorf("unable to get vpc id from instance metadata: %v", err), true
-		}
-		return instanceNotLaunchedInVPCError, false
-	}
-
-	subnetID, err := agent.ec2MetadataClient.SubnetID(mac)
-	if err != nil {
-		return fmt.Errorf("unable to get subnet id from instance metadata: %v", err), false
-	}
-	agent.vpc = vpcID
-	agent.subnet = subnetID
-	agent.mac = mac
-	return nil, false
-}
-
-// isInstanceLaunchedInVPC returns false when the awserr returned is an EC2MetadataError
-// when querying the vpc id from instance metadata
-func isInstanceLaunchedInVPC(err error) bool {
-	if aerr, ok := err.(awserr.Error); ok &&
-		aerr.Code() == "EC2MetadataError" {
-		return false
-	}
-	return true
-}
-
 // verifyCNIPluginsCapabilities returns an error if there's an error querying
 // capabilities or if the required capability is absent from the capabilities
 // of the following plugins:
@@ -180,16 +142,6 @@ func (agent *ecsAgent) startENIWatcher(state dockerstate.TaskEngineState, stateC
 	}
 	go eniWatcher.Start()
 	return nil
-}
-
-func contains(capabilities []string, capability string) bool {
-	for _, cap := range capabilities {
-		if cap == capability {
-			return true
-		}
-	}
-
-	return false
 }
 
 // initializeResourceFields exists mainly for testing doStart() to use mock Control
