+++ agent/config/config.go
@@ -134,6 +134,8 @@ const (
 	//Known cached image names
 	CachedImageNamePauseContainer = "amazon/amazon-ecs-pause:0.1.0"
 	CachedImageNameAgentContainer = "amazon/amazon-ecs-agent:latest"
+	// Known cached windows pause image
+	CachedImageNamePauseContainerWindows = "amazonaws.com/ecs/pause-windows:latest"
 
 	// DefaultNvidiaRuntime is the name of the runtime to pass Nvidia GPUs to containers
 	DefaultNvidiaRuntime = "nvidia"
